extras/m2m-agent/server/requirements.txt
google-adk>=1.25.0
httpx>=0.27.0
fastapi>=0.110.0
uvicorn>=0.29.0
